refactor: Spotlights
